GeoChart footer small bug fixes Start of cleanup Deleting old checkboxlist checkbox-list-enhanced renamed to checkbox-list rush update Reworded the TODOs in utilities so that it lists more nicely in the TODO manager, otherwise it wasn't showing any info Checkboxlist useEffect logic improved
