+++ packages/geoview-core/public/templates/package-footer-panel-geochart.html
@@ -38,7 +38,7 @@
     </div>
 
     <hr />
-    <div id="UIM10" class="geoview-map" data-lang="en" data-config-url="./configs/package-footer-panel-geochart-config.json"></div>
+    <div id="UIM10" class="geoview-map" data-lang="en" data-config-url="./configs/package-footer-panel-geochart-map1-config.json"></div>
     <hr />
     <div id="UIM11" class="geoview-map" data-lang="fr" data-config-url="./configs/package-footer-panel-geochart-map2-config.json"></div>
     <hr />
