+++ packages/geoview-core/src/ui/index.ts
@@ -20,9 +20,7 @@ export * from './icon-button/icon-button';
 export * from './icons';
 export * from './layout';
 export * from './linear-progress/linear-progress';
-export * from './list/checkbox-list/checkbox-list-api';
 export * from './list/checkbox-list/checkbox-list';
-export * from './list/checkbox-list/checkbox-list-enhanced';
 export * from './list';
 export * from './menu';
 export * from './modal';
