+++ packages/geoview-core/src/ui/panel/panel-api.ts
@@ -2,9 +2,6 @@ import { createElement, ReactNode } from 'react';
 
 import { api } from '@/app';
 import { EVENT_NAMES } from '@/api/events/event-types';
-
-import { CheckboxListAPI } from '../list/checkbox-list/checkbox-list-api';
-
 import { PanelPayload, TypeActionButton } from '@/api/events/payloads';
 import { generateId } from '@/core/utils/utilities';
 import { PanelStyles, TypePanelProps } from './panel-types';
@@ -42,8 +39,6 @@ export class PanelApi {
 
   // reference to the map id
   mapId: string;
-
-  checkboxListAPI?: CheckboxListAPI;
 
   panelStyles?: PanelStyles;
 
@@ -166,19 +161,6 @@ export class PanelApi {
     );
 
     return this;
-  };
-
-  /**
-   * Create a check list that can be used as a content
-   *
-   * @param {String[]} elements of the check list the content to update to
-   *
-   * @returns {CheckboxList} the check list
-   */
-  attachCheckBoxList = (listItems: string[], multiselectFlag?: boolean, checkedItems?: number[]): void => {
-    if (this.checkboxListAPI) delete this.checkboxListAPI;
-    this.checkboxListAPI = new CheckboxListAPI(listItems, multiselectFlag, checkedItems);
-    this.changeContent(this.checkboxListAPI.CheckboxList);
   };
 
   /**
